+++ src/includes/_header.pug
@@ -44,7 +44,6 @@ header.header.closed
               button.main-row__utils__catalog-button 
                 include ../svgs/catalog.svg
                 | Catalog  
-                span.ripple
 
               div(class='main-row__utils-buttons-container')
                 button(class=`icon-button button-compare`,data-label='0')
