+++ src/styles/__ripple.scss
@@ -1,0 +1,21 @@
+.ripple {
+  background: currentColor;
+  opacity: 0.6;
+  position: absolute;
+  display: none;
+  border-radius: 50%;
+  transform: translate(-50%, -50%);
+  animation: animate 0.8s linear infinite;
+}
+
+@keyframes animate {
+  0% {
+    width: 0px;
+    height: 0px;
+  }
+  100% {
+    width: 420px;
+    height: 420px;
+    opacity: 0;
+  }
+}
